requirements.txt
Mezzanine==4.2.0
psycopg2==2.6.2
backports.csv==1.0.2
